+++ src/app/layout/mobile-view/components/Footer.tsx
