+++ src/app/layout/mobile-view/components/NavItem.tsx
