+++ src/app/layout/mobile-view/index.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom'
-import Header from './Header'
-import Footer from './Footer'
+import Header from './components/Header'
+import Footer from './components/Footer'
 
 const MobileView = () => {
   return (
