+++ src/app/layout/mobile-view/type/INavItemProps.ts
